WIP: starts to remove token in inapp communication
* starts to remove inapp http communication
* starts to transition to app-role
